Fix BrowserNetworking compilation, returned back RersponseType from swift, to be bale to use Decodable, because Kotlin impl only creates a reference type and you can't extend it with Decodable protocol because it is a reference type.
Now ResponseType is not associated with the Ednpoint which is not good, because you can use wrong response type for the endpoint.

But it could be solved by moving http requests completely to the Kotlin and it will allow to remove Alamofire, which is actually a great library, but it can't be used on Android in the future for this business logic.