Attempt to fix threading issue which is accessing a kotlin object from a different thread from which it was createad. So, e.g. it was created on a main queue but we start use Endpoint type to create a platform URL request when it happens already in background queue inside one of Async APIs like Rx or Combine.
It  should work, but there is a compilation error about using Generic in Endpoint which describes the associated ServerDescription. So, the issue is that Kotlin returns Endpoint type generic as a base interface/base class and on Swift side it can't convert it or recognize it as a concrete Server type which we used on call site in Swift.

Basically, We use specific Server type on call site in Swift, but Kotlin returns base interface and Swift can't compile because it is a different type and base interface for ServerDescription or even abstract base class can't be recognized by Swift as a similar class. Probably it is because we can't up-cast.