Replace Endpoint type in all places (#32)
* Use Kotlin Endpoint in HttpKit. now need to fix the use places in other frameworks like BrowserNetworking

* Fix ReactiveHttpKit to use Kotlin replacement types for ResponseType and for Endpoint

* Fix BrowserNetworking compilation, returned back RersponseType from swift, to be bale to use Decodable, because Kotlin impl only creates a reference type and you can't extend it with Decodable protocol because it is a reference type.

Now ResponseType is not associated with the Ednpoint which is not good, because you can use wrong response type for the endpoint.

But it could be solved by moving http requests completely to the Kotlin and it will allow to remove Alamofire, which is actually a great library, but it can't be used on Android in the future for this business logic.

* Attempt to fix threading issue which is accessing a kotlin object from a different thread from which it was createad. So, e.g. it was created on a main queue but we start use Endpoint type to create a platform URL request when it happens already in background queue inside one of Async APIs like Rx or Combine.

It  should work, but there is a compilation error about using Generic in Endpoint which describes the associated ServerDescription. So, the issue is that Kotlin returns Endpoint type generic as a base interface/base class and on Swift side it can't convert it or recognize it as a concrete Server type which we used on call site in Swift.

Basically, We use specific Server type on call site in Swift, but Kotlin returns base interface and Swift can't compile because it is a different type and base interface for ServerDescription or even abstract base class can't be recognized by Swift as a similar class. Probably it is because we can't up-cast.

* Remove generic server type from Kotlin Endpoint implementation. It also contains some crash but a different one. But I will try to keep the generic in Endpoint (need to try up-cast on Swift side)

* Revert "Remove generic server type from Kotlin Endpoint implementation. It also contains some crash but a different one. But I will try to keep the generic in Endpoint (need to try up-cast on Swift side)"

This reverts commit 06503551eb325c1af040b80c86aff7fa8f6527eb.

* Improve freeze property of the Kotlin implementations. Still perhaps the Endpoint type raises the exception even if it is accessed on the same Thread.

* Add ability to debug Kotlin code in Xcode using  https://github.com/touchlab/xcode-kotlin

Attempt to fix the immutability issue or threading issue

* Fix all kotlin exceptions/crashes. But it is a bad fix. Because Ktor HTTP library has bad documentation or not fully support kotlin native and multiplatform it is not convinietn to always call kotlin code using main thread.

* Add more docs for Kotlin types

* Implement DomainName type in Kotlin, also, ported Punycoding function to not depend on 3rd party libraries